Refactor / organize files
