+++ android/local.properties
@@ -1,0 +1,2 @@
+sdk.dir=/home/fractal/Android/Sdk
+flutter.sdk=/home/fractal/fvm/versions/beta
