+++ lib/src/models/ping_error.dart
@@ -1,0 +1,6 @@
+/// Error code
+enum PingError {
+  RequestTimedOut,
+  UnknownHost,
+  Unknown,
+}
